requirements.txt
accelerate==1.10.1
deepspeed==0.17.6
diffusers==0.35.1
einops==0.8.1
gradio==5.47.0
gradio-client==1.13.2
huggingface-hub==0.35.1
kornia==0.8.1
kornia-rs==0.1.9
numpy==2.2.6
omegaconf==2.3.0
open-clip-torch==2.24.0
opencv-python==4.12.0.88
pillow==11.3.0
safetensors==0.6.2
timm==1.0.20
tokenizers==0.22.1
torch==2.8.0
torchdiffeq==0.2.5
torchsde==0.2.6
torchvision==0.23.0
tqdm==4.67.1
transformers==4.57.0
triton==3.4.0
flash_attn==2.8.3
xformers==0.0.32.post2
